move seed data out of startup
