+++ src/Host/SeedData.cs
@@ -1,0 +1,73 @@
+﻿using System;
+using System.Linq;
+using Host.Configuration;
+using IdentityServer4.EntityFramework.Interfaces;
+using IdentityServer4.EntityFramework.Mappers;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Host
+{
+    public class SeedData
+    {
+        public static void EnsureSeedData(IServiceProvider serviceProvider)
+        {
+            using (var scope = serviceProvider.GetRequiredService<IServiceScopeFactory>().CreateScope())
+            {
+                using (var context = scope.ServiceProvider.GetService<IConfigurationDbContext>())
+                {
+                    EnsureSeedData(context);
+                }
+            }
+        }
+
+        private static void EnsureSeedData(IConfigurationDbContext context)
+        {
+            Console.WriteLine("Seeding database...");
+
+            if (!context.Clients.Any())
+            {
+                Console.WriteLine("Clients being populated");
+                foreach (var client in Clients.Get().ToList())
+                {
+                    context.Clients.Add(client.ToEntity());
+                }
+                context.SaveChanges();
+            }
+            else
+            {
+                Console.WriteLine("Clients already populated");
+            }
+
+            if (!context.IdentityResources.Any())
+            {
+                Console.WriteLine("IdentityResources being populated");
+                foreach (var resource in Resources.GetIdentityResources().ToList())
+                {
+                    context.IdentityResources.Add(resource.ToEntity());
+                }
+                context.SaveChanges();
+            }
+            else
+            {
+                Console.WriteLine("IdentityResources already populated");
+            }
+
+            if (!context.ApiResources.Any())
+            {
+                Console.WriteLine("ApiResources being populated");
+                foreach (var resource in Resources.GetApiResources().ToList())
+                {
+                    context.ApiResources.Add(resource.ToEntity());
+                }
+                context.SaveChanges();
+            }
+            else
+            {
+                Console.WriteLine("ApiResources already populated");
+            }
+
+            Console.WriteLine("Done seeding database.");
+            Console.WriteLine();
+        }
+    }
+}
