+++ src/Host/Startup.cs
@@ -9,10 +9,6 @@ using IdentityServer4.Quickstart.UI;
 using Microsoft.Extensions.Configuration;
 using Microsoft.EntityFrameworkCore;
 using System.Reflection;
-using Host.Configuration;
-using System.Linq;
-using IdentityServer4.EntityFramework.Mappers;
-using IdentityServer4.EntityFramework.Interfaces;
 using Microsoft.AspNetCore.Hosting;
 
 namespace Host
@@ -63,12 +59,6 @@ namespace Host
 
             services.AddMvc();
 
-            // only want this during testing
-            if (_env.IsDevelopment())
-            {
-                //EnsureSeedData(services);
-            }
-
             return services.BuildServiceProvider(validateScopes: true);
         }
 
@@ -81,48 +71,5 @@ namespace Host
             app.UseStaticFiles();
             app.UseMvcWithDefaultRoute();
         }
-
-        private static void EnsureSeedData(IServiceCollection services)
-        {
-            var sp = services.BuildServiceProvider();
-            using (var scope = sp.GetRequiredService<IServiceScopeFactory>().CreateScope())
-            {
-                using (var context = scope.ServiceProvider.GetService<IConfigurationDbContext>())
-                {
-                    EnsureSeedData(context);
-                }
-            }
-        }
-
-        private static void EnsureSeedData(IConfigurationDbContext context)
-        {
-            if (!context.Clients.Any())
-            {
-                foreach (var client in Clients.Get().ToList())
-                {
-                    context.Clients.Add(client.ToEntity());
-                }
-                context.SaveChanges();
-            }
-
-            if (!context.IdentityResources.Any())
-            {
-                foreach (var resource in Resources.GetIdentityResources().ToList())
-                {
-                    context.IdentityResources.Add(resource.ToEntity());
-                }
-                context.SaveChanges();
-            }
-
-            if (!context.ApiResources.Any())
-            {
-                foreach (var resource in Resources.GetApiResources().ToList())
-                {
-                    context.ApiResources.Add(resource.ToEntity());
-                }
-                context.SaveChanges();
-            }
-        }
-
     }
 }
